backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
websockets>=12,<16
httpx>=0.27,<1
pydantic>=2.11,<3
python-dotenv>=1.0,<2
